added the tetris file
input the server/localhost code in the """ part, and make it output the rt and ps variables
                Here does the amazing stuff
                """ 
ps - the position the block will be put in (0;9)
rt - the rotation of the block (0;3)
sight - the state of the field (10 values between 0 and about 20, displays the y coordinate of the highest block in that position)

T.gb is the "Legally" downloaded version of tetris 1987 for the nintendo gameboy.
New text Document.py is the file that runs the server
Note that to run it, it requires a library, called pyboy(can be found on github. im too lazy to look for it)
Thats about it
Pasaki mammai čau no manis, dāvi!